hashery_dutchie_loader/requirements.txt
flask==3.0.2
requests==2.32.3
pandas==2.2.2
google-cloud-bigquery==3.25.0
google-cloud-secret-manager==2.20.0
pyarrow==17.0.0
